+++ client/src/pages/Lessons/ContentHeader.tsx
