+++ client/src/pages/Students/Content.tsx
@@ -1,0 +1,3 @@
+export function StudentsContent() {
+  return <></>;
+}
