--- client/src/pages/Students/StudentsContent.tsx
@@ -1,3 +1,0 @@
-export function StudentsPage() {
-  return <></>;
-}
