+++ client/src/pages/Students/index.tsx
@@ -1,1 +1,1 @@
-export { StudentsPage } from './StudentsPage';
+export { StudentsPage } from './Page';
