+++ client/src/pages/Timetable/ContentCard.tsx
