+++ client/src/pages/Timetable/Page.tsx
@@ -3,8 +3,8 @@ import Stack from '@mui/material/Stack';
 import Box from '@mui/material/Box';
 import CircularProgress from '@mui/material/CircularProgress';
 import useMediaQuery from '@mui/material/useMediaQuery';
-import { DateSwitcher } from './TimetablePageHeader';
-import { DayColumns } from './DayColumns';
+import { DateSwitcher } from './PageHeader';
+import { DayColumns } from './Content';
 import { useGetLessonsQuery } from '../../shared/api/lessonApi';
 import { useAppDispatch } from '../../shared/hooks/useAppDispatch';
 import { setPageTitle } from '../../shared/reducers/appMenuSlice';
